Add Sonoff RFbridge payloads
The Sonoff Tasmota RFBridge will post a message like "message loxone/RFBridge1/tele/RESULT {"Time":"2020-10-13T15:44:55","RfReceived":{"Sync":12540,"Low":440,"High":1250,"Data":"D5K1BE","RfKey":"None"}}" on the MQTT controller when an RF action is received.  This is not understandable by Loxone so we convert this to "send datagram loxone/D5K1BE/cmnd/POWER=1" this way we can filter in loxone on the device name that sends the package to the Rfbridge. make sure you configure a on/off switch in loxone and add an automatic off counter to the input. as the RFBridge will only send ON commands and no OFF commands.